Successfully generated Ranks for first six events
Ranks for the first six events were not saved to a file. Now there's a debug method! One of the options lets you generate a rank file based on a quals csv (provided it fits the format; modularity might be a bit hard later with the current approach but time to learn about dictionaries!) and the ones that I needed are here.